Correcting language around p-values
p-values is the likelihood that if the test groups were drawn from the same distribution, one would detect a difference of this magnitude.  It is NOT the chance that the test is wrong.

This is much different than saying "there’s at most an $\alpha$ chance that the difference arose from random chance. In 95 out of 100 parallel universes, your paper found a difference that actually exists."  In fact, the opposite is what has occurred.  Our test has told us that, where the groups actually identical, we'd still call them different in 5 of 100 parallel universes.  We have no idea what % chance that we're in one of those unlucky 5 universes!  Instead all we can say is "Look, if these groups were equal we'd only get this result 5% of the time, it seems unlikely we're in that 5%."

Consider a test that simply compares one group to itself.  You and I know the groups are identical.  But to a blind tester who returns the results p < 0.05, we know it is 100% due to random chance (not "at most 5%").

This is a big reason why replication studies are gaining emphasis in many fields.  If the discovered difference truly exists, multiple groups are unlikely to each land in one of those 5 unlucky universes.